Fixed bug where the program counter would not increment in JumpIf* instructions if the condition was false.
* Changed `instruction_size` for the `JumpIf*` instructions to return
their correct size (3) instead of (0).
What happend was that the program counter would **not** increment
if the `JumpIf*` instruction's condition was not met. This resulted in
an infinite loop.
* Changed the `JumpIf*` implementations to set the `pc` to
`location.wrapping_sub(opcode.instruction_size())` - this means that the
`pc` will increment correctly to the next location.
* Changed the incrementation of the program counter (`pc`) to use
`wrapping_add` - this means that if the `JumpIf*` instructions jumped
to the beginning (0), the `pc` will correctly increment.